services/frontend-service/requirements.txt
fastapi[standard]
uvicorn[standard]
jinja2>=3.1.2
httpx>=0.27.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Service Discovery
python-consul2>=0.1.5

# Monitoring and Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-exporter-otlp>=1.21.0
structlog>=23.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
